Moved Location to typeclasses
