+++ src/structure.js
@@ -10,6 +10,7 @@ import desugar from './desugar';
 import Microstate from './microstate';
 import { collapse } from './typeclasses/collapse';
 import logTree from './utils/log-tree';
+import { stateAt, childrenAt } from './typeclasses/location';
 
 const { assign } = Object;
 
@@ -45,53 +46,6 @@ export function collapseState(tree) {
   let truncated = truncate(node => node.isSimple, tree);
   return collapse(map(node => node.state, truncated));
 }
-
-const Location = type(class Location {
-  stateAt(Type, instance, value) {
-    return this(Type.prototype).stateAt(instance, value);
-  }
-  childrenAt(Type, value) {
-    return this(Type.prototype).childrenAt(Type, value);
-  }
-});
-
-const { stateAt, childrenAt } = Location.prototype;
-
-Location.instance(Object, {
-  stateAt(instance, value) {
-    if (value) {
-      return append(instance, value);
-    } else {
-      return instance;
-    }
-  },
-
-  childrenAt(Type, value) {
-    return $(new Type())
-      .map(desugar)
-      .filter(({ value }) => !!value && value.call)
-      .valueOf();
-  }
-});
-
-Location.instance(types.Object, {
-  stateAt: _ => ({}),
-  childrenAt(Type, value) {
-    let { T } = params(Type);
-    if (T !== any) {
-      return map(_ => T, value);
-    } else {
-      return Location.for(Object).childrenAt(Type, value);
-    }
-  }
-});
-
-Location.instance(types.Array, {
-  stateAt: _ => [],
-  childrenAt(...args) {
-    return Location.for(types.Object.prototype).childrenAt(...args);
-  }
-});
 
 function truncate(fn, tree) {
   return flatMap(node => {
